Add `generate_prelude` and `generate_support_types`.
