+++ codegen/src/generator/json.rs
@@ -10,24 +10,26 @@ impl GenerateProtocol for JsonGenerator {
         service.operations.values().map(|operation| {
             let output_type = operation.output_shape_or("()");
 
-            format!("{documentation}
-pub fn {method_name}(&mut self, input: &{input_type}) -> Result<{output_type}> {{
-    let encoded = serde_json::to_string(input).unwrap();
-    let mut request = SignedRequest::new(\"{http_method}\", \"{endpoint_prefix}\", &self.region, \"{request_uri}\");
-    request.set_content_type(\"application/x-amz-json-1.0\".to_owned());
-    request.add_header(\"x-amz-target\", \"{target_prefix}.{name}\");
-    request.set_payload(Some(encoded.as_bytes()));
-    let mut result = request.sign_and_execute(try!(self.credentials_provider.credentials()));
-    let status = result.status.to_u16();
-    let mut body = String::new();
-    result.read_to_string(&mut body).unwrap();
-    match status {{
-        200 => {{
-            {ok_response}
-        }}
-        _ => Err(parse_error(&body)),
-    }}
-}}",
+            format!("
+                {documentation}
+                pub fn {method_name}(&mut self, input: &{input_type}) -> Result<{output_type}> {{
+                    let encoded = serde_json::to_string(input).unwrap();
+                    let mut request = SignedRequest::new(\"{http_method}\", \"{endpoint_prefix}\", &self.region, \"{request_uri}\");
+                    request.set_content_type(\"application/x-amz-json-1.0\".to_owned());
+                    request.add_header(\"x-amz-target\", \"{target_prefix}.{name}\");
+                    request.set_payload(Some(encoded.as_bytes()));
+                    let mut result = request.sign_and_execute(try!(self.credentials_provider.credentials()));
+                    let status = result.status.to_u16();
+                    let mut body = String::new();
+                    result.read_to_string(&mut body).unwrap();
+                    match status {{
+                        200 => {{
+                            {ok_response}
+                        }}
+                        _ => Err(parse_error(&body)),
+                    }}
+                }}
+                ",
                 documentation = generate_documentation(operation).unwrap_or("".to_owned()),
                 endpoint_prefix = service.metadata.endpoint_prefix,
                 http_method = operation.http.method,
@@ -41,6 +43,56 @@ pub fn {method_name}(&mut self, input: &{input_type}) -> Result<{output_type}> {
             )
         }).collect::<Vec<String>>().join("\n")
     }
+
+    fn generate_prelude(&self, service: &Service) -> String {
+        format!(
+            "use std::io::Read;
+            use std::result;
+
+            use serde_json;
+
+            use credentials::ProvideAWSCredentials;
+            use error::AWSError;
+            use regions::Region;
+            use signature::SignedRequest;
+
+            /// An error returned by the {service_name} API.
+            #[derive(Debug, Default, Deserialize)]
+            pub struct {error_type_name} {{
+                __type: String,
+                message: String,
+            }}
+
+            /// On success: `T`. On error: `{error_type_name}`.
+            pub type Result<T> = result::Result<T, {error_type_name}>;
+
+            impl From<AWSError> for {error_type_name} {{
+                fn from(err: AWSError) -> Self {{
+                    let AWSError(message) = err;
+
+                    {error_type_name} {{
+                        __type: \"Unknown\".to_string(),
+                        message: message.to_string(),
+                    }}
+                }}
+            }}
+
+            fn parse_error(body: &str) -> {error_type_name} {{
+                if let Ok(decoded) = serde_json::from_str::<{error_type_name}>(&body) {{
+                    decoded
+                }} else {{
+                    {error_type_name} {{
+                        __type: \"DecodeError\".to_string(),
+                        message: body.to_string(),
+                    }}
+                }}
+            }}
+            ",
+            error_type_name = service.error_type_name(),
+            service_name = service.service_type_name(),
+        )
+
+    }
 }
 
 fn generate_documentation(operation: &Operation) -> Option<String> {
@@ -51,10 +103,7 @@ fn generate_documentation(operation: &Operation) -> Option<String> {
 
 fn generate_ok_response(operation: &Operation, output_type: &str) -> String {
     if operation.output.is_some() {
-        format!(
-            "Ok(serde_json::from_str::<{}>(&body).unwrap())",
-            output_type,
-        )
+        format!("Ok(serde_json::from_str::<{}>(&body).unwrap())", output_type)
     } else {
         "Ok(())".to_owned()
     }
